src/requirements.txt
gspread
oauth2client            # for Google Sheets auth
openpyxl
networkx
neo4j
fastapi                # later, for our API layer
uvicorn
llama-index 
llama-index-graph-stores-nebula
llama-index-graph-stores-neo4j
llama-index-embeddings-huggingface
pyvis
watchdog
python-dotenv
typer